move stopwatch to proxy and do not export it
